fix: rename analytics icon file name to statatics to prevent ad blocking
